feat(compile): per-target runtime archives + toolchain reorg + multi-target CI
Prep para cross-compile (desktop/Android/iOS):

- Artifacts per-target: ~/.rts/artifacts/<triple>.a (antes: artifacts.a só host).
  host = archive embutido (sha/decompress); cross = RTS_RUNTIME_OBJECTS_DIR/
  <triple>.a -> runtime-objects/<triple>.a ao lado do exe -> RTS_RUNTIME_ARCHIVE_URL
  (template {target}) -> bail com instrução. build.rs emite RTS_HOST_TARGET;
  lib.rs expõe rt_artifacts_for_target().
- Toolchain reorg: rts-linker/src/toolchain.rs (1157 linhas) -> raiz slim (90) +
  toolchain/{target,paths,discover,download,windows_sdk}.rs (todos <500). API
  pública idêntica via pub use; zero mudança de comportamento.
- CI runtime-archives.yml: matrix por-target publica <triple>.a no release —
  win-x64, linux-x64, linux-arm64 (cross gcc), mac arm64/x64, android-arm64 (NDK)
  e iOS device+sim (build no macOS, SDK via Xcode). Android/iOS best-effort
  (continue-on-error) até o backend de áudio mobile (#1723).
